+++ dataset_generation/dataset.py
@@ -1,6 +1,6 @@
 from os import devnull
 from time import sleep
-from typing import List, Optional
+from typing import List, Optional, Union
 from pathlib import Path
 from subprocess import call
 from json import loads, dumps
@@ -13,17 +13,18 @@ from tdw.py_impact import PyImpact, ObjectInfo, AudioMaterial
 from tdw.output_data import Rigidbodies, Transforms, AudioSources
 from magnebot import ActionStatus
 from magnebot.scene_state import SceneState
-from magnebot.constants import OCCUPANCY_CELL_SIZE
 from magnebot.util import get_data
+from magnebot.constants import OCCUPANCY_CELL_SIZE
 from multimodal_challenge.multimodal_base import MultiModalBase
 from multimodal_challenge.paths import REHEARSAL_DIRECTORY, ENV_AUDIO_MATERIALS_PATH, DATASET_DIRECTORY,\
-    OBJECT_INIT_DIRECTORY, SCENE_BOUNDS_DIRECTORY, MAGNEBOT_OCCUPANCY_MAPS_DIRECTORY
+    OBJECT_INIT_DIRECTORY, SCENE_BOUNDS_DIRECTORY
 from multimodal_challenge.util import get_scene_layouts, get_trial_filename
 from multimodal_challenge.multimodal_object_init_data import MultiModalObjectInitData
 from multimodal_challenge.trial import Trial
 from multimodal_challenge.encoder import Encoder
 from multimodal_challenge.dataset.dataset_trial import DatasetTrial
 from multimodal_challenge.dataset.env_audio_materials import EnvAudioMaterials
+from multimodal_challenge.dataset.add_ons.occupancy_map import OccupancyMap
 
 
 class Dataset(MultiModalBase):
@@ -109,10 +110,11 @@ class Dataset(MultiModalBase):
 
     **Result:** A directory dataset files. The dataset has a `random_seeds.npy` file that is used to select random seeds per trial.
 
-    Each trial is saved in a `scene_layout` directory and has two files:
+    Each trial is saved in a `scene_layout` directory and has three files:
     
     1. A .json file of the [`Trial` data](../api/trial.md).
-    2. An audio .wave file.
+    2. An audio .wav audio file.
+    3. The occupancy map as a .npy numpy file.
     
     ```
     D:/multimodal_challenge/
@@ -120,8 +122,10 @@ class Dataset(MultiModalBase):
     ....mm_kitchen_1a_0/  # scene_layout
     ........00000.json
     ........00000.wav
+    ........00000.npy
     ........00001.json
     ........00001.wav
+    ........00001.npy
     ........(etc.)
     ....mm_kitchen_1a_1/
     ```
@@ -144,14 +148,21 @@ class Dataset(MultiModalBase):
     """
     TEMP_AUDIO_PATH: Path = DATASET_DIRECTORY.joinpath("temp.wav")
 
-    def __init__(self, port: int = 1071, random_seed: int = 0):
+    def __init__(self, port: int = 1071, random_seed: int = 0, log: bool = True):
         """
         Create the network socket and bind the socket to the port.
 
         :param port: The port number.
         :param random_seed: The seed for the random number generator.
+        :param log: If True, log each list of commands sent.
         """
-
+        
+        if not DATASET_DIRECTORY.exists():
+            DATASET_DIRECTORY.mkdir(parents=True)
+        self._log: bool = log
+        self._log_path: Path = DATASET_DIRECTORY.joinpath("log.txt")
+        if self._log_path.exists():
+            self._log_path.unlink()
         super().__init__(port=port, random_seed=random_seed, screen_height=128, screen_width=128, skip_frames=0)
         self.communicate([{"$type": "set_render_quality",
                            "render_quality": 0},
@@ -191,6 +202,8 @@ class Dataset(MultiModalBase):
         # which allows us to pause/resume dataset generation without inadvertantly changing it.
         self._random_seeds: np.array = np.array([])
         self._random_seed_index: int = 0
+        # The IDs of the target object and the distractors.
+        self._extra_object_ids: List[int] = list()
 
     def run(self) -> None:
         """
@@ -198,8 +211,6 @@ class Dataset(MultiModalBase):
         """
 
         random_seed_path = DATASET_DIRECTORY.joinpath("random_seeds.npy").resolve()
-        if not DATASET_DIRECTORY.exists():
-            DATASET_DIRECTORY.mkdir(parents=True)
         # Load existing random seeds.
         if random_seed_path.exists():
             self._random_seeds = np.load(str(random_seed_path))
@@ -265,7 +276,7 @@ class Dataset(MultiModalBase):
             for i in range(self.trial_count, len(self.trials)):
                 self.do_trial(output_directory=output_directory)
                 pbar.update(1)
-        # Close the audio thread, stop fmedia, and stop the progress bar.
+        # Stop fmedia from recording.
         finally:
             AudioUtils.stop()
 
@@ -376,6 +387,19 @@ class Dataset(MultiModalBase):
                 i.gravity = True
                 target_object_index = len(object_init_data)
             object_init_data.append(i)
+
+        # Re-create the occupancy map to include the distractors and the target object.
+        occupancy_mapper = OccupancyMap(cell_size=OCCUPANCY_CELL_SIZE)
+        resp = self.communicate(occupancy_mapper.get_initialization_commands())
+        occupancy_mapper.initialized = True
+        occupancy_mapper.on_send(resp=resp)
+        occupancy_mapper.generate()
+        resp = self.communicate(occupancy_mapper.commands)
+        occupancy_mapper.on_send(resp=resp)
+        self.occupancy_map = occupancy_mapper.occupancy_map
+
+        # Update the scene state (just in case something actually moved).
+        state = SceneState(resp=resp)
         # Create the trial.
         trial = Trial(object_init_data=object_init_data,
                       target_object_index=target_object_index,
@@ -385,6 +409,8 @@ class Dataset(MultiModalBase):
         filename = get_trial_filename(self.trial_count)
         # Save the trial.
         output_directory.joinpath(f"{filename}.json").write_text(dumps(trial, cls=Encoder), encoding="utf-8")
+        # Save the occupancy map.
+        np.save(str(output_directory.joinpath(filename).resolve()), self.occupancy_map)
         # Use ffmpeg to remove the initial silence.
         with open(devnull, "w+") as f:
             call(["ffmpeg", "-i", str(Dataset.TEMP_AUDIO_PATH.resolve()),
@@ -413,10 +439,18 @@ class Dataset(MultiModalBase):
             o_id, o_commands = MultiModalObjectInitData(**o).get_commands()
             self._object_init_commands[o_id] = o_commands
         # Add the target object.
-        self.target_object_id, target_object_commands = self.trials[self.trial_count].init_data.get_commands()
+        self.target_object_id, target_object_commands = self.trials[self.trial_count].target_object.get_commands()
+        self._extra_object_ids.clear()
+        self._extra_object_ids.append(self.target_object_id)
         self._object_init_commands[self.target_object_id] = target_object_commands
+        # Add the distractor objects.
+        for distractor in self.trials[self.trial_count].distractors:
+            o_id, o_commands = distractor.get_commands()
+            self._extra_object_ids.append(o_id)
+            self._object_init_commands[o_id] = o_commands
         # We need every frame for audio recording, but not right now, so let's speed things up.
         self._skip_frames = 10
+        self._scene_bounds = loads(SCENE_BOUNDS_DIRECTORY.joinpath(f"{scene[:-1]}.json").read_text())
         # Initialize the scene.
         super().init_scene(scene=scene, layout=layout)
         # Turn the Magnebot by a random angle.
@@ -442,6 +476,17 @@ class Dataset(MultiModalBase):
         Dataset.PY_IMPACT.reset(initial_amp=Dataset.INITIAL_AMP)
         return ActionStatus.success
 
+    def communicate(self, commands: Union[dict, List[dict]]) -> List[bytes]:
+        # Log the message.
+        if self._log:
+            if isinstance(commands, list):
+                msg = dumps(commands)
+            else:
+                msg = dumps([commands])
+            with self._log_path.open("at", encoding="utf-8") as f:
+                f.write(msg + "\n")
+        return super().communicate(commands=commands)
+
     def _cache_static_data(self, resp: List[bytes]) -> None:
         super()._cache_static_data(resp=resp)
         self.env_id = self.get_unique_id()
@@ -466,25 +511,7 @@ class Dataset(MultiModalBase):
                 {"$type": "add_environ_audio_sensor"}]
 
     def _get_magnebot_position(self) -> np.array:
-
-        scene_bounds = loads(SCENE_BOUNDS_DIRECTORY.joinpath(f"{self.scene[:-1]}.json").read_text())
-        # Get all free occupancy map positions.
-        target_object_position = TDWUtils.vector3_to_array(self.trials[self.trial_count].init_data.position)
-
-        magnebot_occupancy_map = np.load(MAGNEBOT_OCCUPANCY_MAPS_DIRECTORY.joinpath(f"{self.scene}_{self.layout}.npy"))
-        magnebot_positions: List[np.array] = list()
-        for ix, iy in np.ndindex(magnebot_occupancy_map.shape):
-            if magnebot_occupancy_map[ix][iy] != 0:
-                continue
-            px = scene_bounds["x_min"] + (ix * OCCUPANCY_CELL_SIZE)
-            pz = scene_bounds["z_min"] + (iy * OCCUPANCY_CELL_SIZE)
-            # Ignore positions close to the target object.
-            p = np.array([px, 0, pz])
-            if np.linalg.norm(p - target_object_position) < OCCUPANCY_CELL_SIZE * 1.1:
-                continue
-            magnebot_positions.append(p)
-        # Pick a random position.
-        return magnebot_positions[self._rng.randint(0, len(magnebot_positions))]
+        return TDWUtils.vector3_to_array(self.trials[self.trial_count].magnebot_position)
 
     def _listen_for_audio(self) -> None:
         """
@@ -534,6 +561,7 @@ if __name__ == "__main__":
     from argparse import ArgumentParser
     parser = ArgumentParser()
     parser.add_argument("--random_seed", type=int, default=0, help="The total number of trials.")
+    parser.add_argument("--log", action="store_true", help="Log all commands sent to the build.")
     args = parser.parse_args()
-    dataset_generator = Dataset(random_seed=args.random_seed)
+    dataset_generator = Dataset(random_seed=args.random_seed, log=args.log)
     dataset_generator.run()