+++ multimodal_challenge/dataset/dataset_trial.py
@@ -1,4 +1,4 @@
-from typing import Dict
+from typing import Dict, List, Union
 from multimodal_challenge.multimodal_object_init_data import MultiModalObjectInitData
 
 
@@ -7,27 +7,41 @@ class DatasetTrial:
     Parameters for defining a trial for dataset generation.
     """
 
-    def __init__(self, init_data: MultiModalObjectInitData, force: Dict[str, float], position: Dict[str, float]):
+    def __init__(self, target_object: MultiModalObjectInitData, force: Dict[str, float],
+                 magnebot_position: Dict[str, float],
+                 target_object_position: Dict[str, float],
+                 distractors: List[Union[dict, MultiModalObjectInitData]]):
         """
-        :param init_data: A [`MultiModalObjectInitData` object](multimodal_object_init_data.md).
-        :param force: The initial force of the object as a Vector3 dictionary.
-        :param position: The position of the object after it falls. This is used to set a valid initial Magnebot pose.
+        :param target_object: [`MultiModalObjectInitData` initialization data](multimodal_object_init_data.md) for the target object.
+        :param force: The initial force of the target object as a Vector3 dictionary.
+        :param magnebot_position: The initial position of the Magnebot.
+        :param target_object_position: The final position of the target object.
+        :param distractors: Initialization data for the distractor objects.
         """
 
         # Load the drop parameters from a dictionary.
-        if isinstance(init_data, dict):
-            init_data: dict
+        if isinstance(target_object, dict):
+            target_object: dict
             """:field
-            A [`MultiModalObjectInitData` object](multimodal_object_init_data.md) for the dropped object.
+            target_object: [`MultiModalObjectInitData` initialization data](multimodal_object_init_data.md) for the target object.
             """
-            self.init_data = MultiModalObjectInitData(**init_data)
+            self.target_object = MultiModalObjectInitData(**target_object)
         else:
-            self.init_data: MultiModalObjectInitData = init_data
+            self.target_object: MultiModalObjectInitData = target_object
         """:field
-        The initial force of the dropped object object as a Vector3 dictionary.
+        The initial force of the target object as a Vector3 dictionary.
         """
         self.force: Dict[str, float] = force
         """:field
-        The position of the object after it falls. This is used to set a valid initial Magnebot pose.
+        The initial position of the Magnebot.
         """
-        self.position: Dict[str, float] = position
+        self.magnebot_position: Dict[str, float] = magnebot_position
+        """:field
+        The final position of the target object.
+        """
+        self.target_object_position: Dict[str, float] = target_object_position
+        """:field
+        Initialization data for the distractor objects.
+        """
+        self.distractors: List[MultiModalObjectInitData] = [d if isinstance(d, MultiModalObjectInitData) else
+                                                            MultiModalObjectInitData(**d) for d in distractors]