+++ doc/changelog.md
@@ -1,3 +1,27 @@
+# 0.4.1
+
+- Required version of TDW is now handled by the Magnebot API. Currently, the required version is 1.8.24.0
+- Required version of Magnebot: 1.3.0
+
+# 0.4.0
+
+- **Added distractor objects.** Objects are randomly placed on the floor during rehearsal.py
+  - (Backend) rehearsal.py sets the initial Magnebot position based on where the objects (distractors and target object) have fallen
+  - (Backend) `DatasetTrial` now includes initialization data for the distractor objects and the initial Magnebot position
+  - (Backend) Added data file `distractor_objects.txt`
+- The target object is now dropped onto a free space of the occupancy map instead of a drop zone
+- Added API documentation for Magnebot arm articulation actions (although they have always been usable in this API)
+- Improved version checking for the `tdw` pip module, the `magnebot` pip module, and the TDW build
+- (Backend): Added `AddOn` and `OccupancyMap` add-ons. These scripts are from a future version of TDW (1.9.0).
+- (Backend): Removed the old `OccupancyMapper` and created a new one in `dataset_generation/occupancy_mapper.py`. This script generates far more accurate occupancy maps using the new add-ons.
+  - Removed occupancy map generation from `init_data.py`
+- (Backend): Added optional parameter `log` to the `Dataset` constructor; if True, log messages.
+- (Backend): **Removed drop zones** (removed the Python class, all references to drop zones in the documentation, and the cached drop zone positions)
+- (Backend): `dataset.py` now saves occupancy maps as .npy files, which include the position of the target object and the distractor objects
+- (Backend): It's now possible to resume `rehearsal.py` rather than having to restart from the beginning
+- (Backend): Added `util/add_models.py` to update the model librarian json file
+- (Backend): Added `packaging` as a required pip module
+
 # 0.3.9
 
 - Required version of TDW: 1.8.21.0